fix: explain sandbox-blocked approval access
