+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![cfg(target_os = "macos")]
 
+mod approval_service;
 pub mod brew_cask_policy;
 mod cli;
 mod gpg;
@@ -9,6 +10,11 @@ mod secrets;
 
 pub use cli::{run, run_scanner_terminal, run_terminal};
 pub use gpg::run_git_program;
+
+#[doc(hidden)]
+pub fn approval_service_unavailable_message(service: &std::ffi::CStr) -> &'static str {
+    approval_service::unavailable_message(service)
+}
 
 pub const MENU_HELPER_CODE_SIGNING_REQUIREMENT: &str = r#"anchor apple generic and certificate leaf[subject.OU] = ZU76A67LGU and identifier "com.automicvault""#;
